VAlidaciones y metodo show

+++ database/migrations/2021_07_06_171423_create_productos_table.php
@@ -23,9 +23,9 @@ class CreateProductosTable extends Migration
             $table->integer('stock');
             $table->string('presentacion');
             $table->string('dimension');
-            $table->float('preciodistribuidor');
-            $table->float('preciocontado');
-            $table->float('preciopublico');
+            $table->double('preciodistribuidor');
+            $table->double('preciocontado');
+            $table->double('preciopublico');
             $table->string('descripcion');
             $table->string('tipoproducto');
             $table->timestamps();
